# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/producers": {
      "post": {
        "operationId": "ProducerController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProducerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produtor criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProducerMaskedResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "producers"
        ]
      },
      "get": {
        "operationId": "ProducerController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 10000,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "name"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 255,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada resumida de produtores (farmsCount/áreas/UFs; sem harvests/crops). First paint da tabela.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProducerListPageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "producers"
        ]
      }
    },
    "/api/v1/producers/search": {
      "get": {
        "operationId": "ProducerController_search",
        "parameters": [
          {
            "name": "document",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 11,
              "maxLength": 18,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Busca por documento (blind index) — retorno hidratado completo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProducerDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "producers"
        ]
      }
    },
    "/api/v1/producers/{id}/esg-compliance": {
      "get": {
        "operationId": "ProducerController_esgCompliance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Parecer socioambiental ESG"
          }
        },
        "tags": [
          "producers"
        ]
      }
    },
    "/api/v1/producers/{id}": {
      "get": {
        "operationId": "ProducerController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhe do produtor com fazendas, safras e culturas (hidratação completa)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProducerDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "producers"
        ]
      },
      "put": {
        "operationId": "ProducerController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProducerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produtor atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProducerDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "producers"
        ]
      },
      "delete": {
        "operationId": "ProducerController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "producers"
        ]
      }
    },
...
    "/api/v1/admin/revalidate/producers/{id}": {
      "post": {
        "operationId": "AdminRevalidateController_producer",
        "parameters": [
          {
            "name": "X-Admin-Token",
            "in": "header",
            "description": "Token administrativo (ADMIN_API_TOKEN)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revalida CNPJ do produtor via BrasilAPI",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevalidateResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Brain Agriculture API",
    "description": "API de gestão de produtores rurais, fazendas, safras e dashboard analítico. Sem autenticação no escopo atual — proteja a rede e use rate limit/CORS em produção.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateProducerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "document": {
            "type": "string",
            "minLength": 11,
            "maxLength": 18
          },
          "farms": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255
                },
                "city": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                },
                "state": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 2
                },
                "totalArea": {
                  "type": "number",
                  "exclusiveMinimum": true,
                  "minimum": 0
                },
                "arableArea": {
                  "type": "number",
                  "minimum": 0
                },
                "vegetationArea": {
                  "type": "number",
                  "minimum": 0
                },
                "harvests": {
                  "maxItems": 10,
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "year": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 10
                      },
                      "crops": {
                        "minItems": 1,
                        "maxItems": 20,
                        "type": "array",
                        "items": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 50
                        }
                      }
                    },
                    "required": [
                      "year",
                      "crops"
                    ],
                    "additionalProperties": false
                  }
                },
                "carNumber": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                }
              },
              "required": [
                "name",
                "city",
                "state",
                "totalArea",
                "arableArea",
                "vegetationArea"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "name",
          "document"
        ],
        "additionalProperties": false
      },
      "ProducerMaskedResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a8f3d1b2-8c9e-4a1b-9f01-123456789abc"
          },
          "name": {
            "type": "string",
            "example": "João da Silva"
          },
          "document": {
            "type": "string",
            "description": "CPF/CNPJ mascarado na resposta",
            "examples": {
              "cpf": {
                "value": "***.982.247-**"
              },
              "cnpj": {
                "value": "**.222.333/0001-**"
              }
            },
            "example": "***.982.247-**"
          }
        },
        "required": [
          "id",
          "name",
          "document"
        ]
      },
      "ProducerListItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a8f3d1b2-8c9e-4a1b-9f01-123456789abc"
          },
          "name": {
            "type": "string",
            "example": "João da Silva"
          },
          "document": {
            "type": "string",
            "description": "CPF/CNPJ mascarado",
            "example": "***.982.247-**"
          },
          "esgStatus": {
            "type": "string",
            "example": "APPROVED"
          },
          "esgCheckedAt": {
            "type": "object",
            "nullable": true,
            "example": "2026-01-15T12:00:00.000Z"
          },
          "documentValidationStatus": {
            "type": "string",
            "enum": [
              "VALIDATED",
              "PENDING_EXTERNAL_VALIDATION",
              "REJECTED"
            ],
            "description": "VALIDATED | PENDING_EXTERNAL_VALIDATION | REJECTED (BrasilAPI CNPJ). Outage → PENDING, nunca positivo silencioso."
          },
          "documentValidationPendingAt": {
            "type": "object",
            "nullable": true
          },
          "documentValidationPendingReason": {
            "type": "object",
            "nullable": true,
            "description": "Motivo da pendência (circuit_open, timeout_or_network, …)"
          },
          "farmsCount": {
            "type": "number",
            "description": "Quantidade de fazendas ativas do produtor",
            "example": 3
          },
          "farmStates": {
            "description": "UFs distintas das fazendas",
            "example": [
              "SP",
              "MG"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalAreaHa": {
            "type": "number",
            "example": 1250.5
          },
          "arableAreaHa": {
            "type": "number",
            "example": 800
          },
          "vegetationAreaHa": {
            "type": "number",
            "example": 450.5
          }
        },
        "required": [
          "id",
          "name",
          "document",
          "esgStatus",
          "documentValidationStatus",
          "farmsCount",
          "farmStates",
          "totalAreaHa",
          "arableAreaHa",
          "vegetationAreaHa"
        ]
      },
      "ProducerListPageResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProducerListItemResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 3334
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "HarvestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "year": {
            "type": "string",
            "example": "2024/2025"
          },
          "status": {
            "type": "string",
            "example": "ACTIVE"
          },
          "crops": {
            "example": [
              "Soja",
              "Milho"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "year",
          "status",
          "crops"
        ]
      },
      "FarmDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "producerId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "example": "SP"
          },
          "totalArea": {
            "type": "number"
          },
          "arableArea": {
            "type": "number"
          },
          "vegetationArea": {
            "type": "number"
          },
          "carNumber": {
            "type": "object",
            "nullable": true
          },
          "carStatus": {
            "type": "object",
            "nullable": true
          },
          "climateRiskScore": {
            "type": "object",
            "nullable": true
          },
          "territorialValidationStatus": {
            "type": "string",
            "enum": [
              "VALIDATED",
              "PENDING_EXTERNAL_VALIDATION",
              "REJECTED"
            ],
            "description": "Validação cidade∈UF via BrasilAPI IBGE"
          },
          "territorialValidationPendingAt": {
            "type": "object",
            "nullable": true
          },
          "territorialValidationPendingReason": {
            "type": "object",
            "nullable": true
          },
          "harvests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HarvestResponseDto"
            }
          }
        },
        "required": [
          "id",
          "producerId",
          "name",
          "city",
          "state",
          "totalArea",
          "arableArea",
          "vegetationArea",
          "territorialValidationStatus",
          "harvests"
        ]
      },
      "ProducerDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "document": {
            "type": "string",
            "description": "CPF/CNPJ mascarado"
          },
          "esgStatus": {
            "type": "string"
          },
          "esgCheckedAt": {
            "type": "object",
            "nullable": true
          },
          "documentValidationStatus": {
            "type": "string",
            "enum": [
              "VALIDATED",
              "PENDING_EXTERNAL_VALIDATION",
              "REJECTED"
            ],
            "description": "Cadastro aceita PENDING; REJECTED/inativo bloqueia no write. Compliance plena exige VALIDATED."
          },
          "documentValidationPendingAt": {
            "type": "object",
            "nullable": true
          },
          "documentValidationPendingReason": {
            "type": "object",
            "nullable": true
          },
          "farms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FarmDetailResponseDto"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "document",
          "esgStatus",
          "documentValidationStatus",
          "farms",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProducerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "document": {
            "type": "string",
            "minLength": 11,
            "maxLength": 18
          }
        },
        "additionalProperties": false
      },
...
      "RevalidateResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "previousStatus": {
            "type": "string",
            "enum": [
              "VALIDATED",
              "PENDING_EXTERNAL_VALIDATION",
              "REJECTED"
            ]
          },
          "newStatus": {
            "type": "string",
            "enum": [
              "VALIDATED",
              "PENDING_EXTERNAL_VALIDATION",
              "REJECTED"
            ]
          },
          "reason": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "previousStatus",
          "newStatus"
        ]
      }
    }
  }
}